+++ example/angular/src/app/radio-group/radio-group-example.component.ts
@@ -1,11 +1,13 @@
 import { Component, OnInit } from '@angular/core';
+import { FormControl, Validators } from "@angular/forms";
 
 @Component({
   selector: 'app-radio-group-example',
   template: `
     <wcs-card style="margin-top: 8px">
       <wcs-card-body>
-        current select value : {{currentValue ? currentValue : 'none'}}
+        <h3>ngModel</h3>
+        Current select value : {{currentValue ? currentValue : 'none'}}
         <wcs-radio-group [(ngModel)]="currentValue">
           <wcs-radio value="1" label="Option 1"></wcs-radio>
           <wcs-radio value="2" label="Option 2"></wcs-radio>
@@ -13,7 +15,22 @@ import { Component, OnInit } from '@angular/core';
           <wcs-radio value="4" label="Option 4"></wcs-radio>
         </wcs-radio-group>
         <br/>
-        <wcs-button class="wcs-primary" mode="stroked" (click)="onClick()">Choose option 4</wcs-button>
+        <wcs-button class="wcs-primary" mode="stroked" (click)="currentValue = '4'">Choose option 4</wcs-button>
+      </wcs-card-body>
+    </wcs-card>
+    <wcs-card style="margin-top: 8px">
+      <wcs-card-body>
+        <h3>formControl</h3>
+        Current select value : {{radioGroup.value || 'none'}}
+        <wcs-radio-group [formControl]="radioGroup">
+          <wcs-radio value="1" label="Option 1"></wcs-radio>
+          <wcs-radio value="2" label="Option 2"></wcs-radio>
+          <wcs-radio value="3" label="Option 3"></wcs-radio>
+          <wcs-radio value="4" label="Option 4"></wcs-radio>
+        </wcs-radio-group>
+        <br/>
+        <wcs-button class="wcs-primary" mode="stroked" (click)="radioGroup.setValue('4')">Choose option 4</wcs-button>
+        <div>radio-group touched ? {{ radioGroup.touched }}</div>
       </wcs-card-body>
     </wcs-card>
   `,
@@ -21,14 +38,13 @@ import { Component, OnInit } from '@angular/core';
 })
 export class RadioGroupExampleComponent implements OnInit {
   currentValue: string;
+  public radioGroup = new FormControl('', []);
+
 
   constructor() {
   }
 
   ngOnInit(): void {
-  }
 
-  onClick() {
-    this.currentValue = '4';
   }
 }